Changes how HiDens data is streamed and emitted.
Previously, data was streamed in chunks, triggered on a timer.
This was less than ideal for several reasons, most prominently
that the data source could only fall behind the actual stream,
never actually catch up.

Further, the data that was emitted was restricted to valid data
channels, meaning the number of channels changes potentially
whenever the configuration changes. This adds a significant
amount of complexity to downstream code, which must handle
a flexible number of data channels. This commit takes a cue from
the Hierlemann group's implemetation, which always sends all
data (regardless of validity), and foists the problem of determing
which data is "valid" onto clients.

This adds some overhead in the network, in that useless data
is sent, and in the files, in that useless data is saved. But
it makes everything else so much less complex, and solves the
real issue that prompted the change: that subsetting the data
correctly based on the current configuration is error-prone and
complicated.